Extract OpenAPI backed plugin helpers
